# Improved ToString() performance.