--- a/tools/LightResults.DevelopBenchmarks/Benchmarks.cs
+++ b/tools/LightResults.DevelopBenchmarks/Benchmarks.cs
@@ -1,31 +1,107 @@
-﻿using System.Collections.Immutable;
-using BenchmarkDotNet.Attributes;
+﻿using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Columns;
+using BenchmarkDotNet.Jobs;
+using LightResults.Common;
 
 namespace LightResults.DevelopBenchmarks;
 
 [MemoryDiagnoser]
+[SimpleJob(RuntimeMoniker.Net80)]
+[HideColumns(Column.Job, Column.Iterations, Column.Error, Column.StdDev, Column.Gen0, Column.Gen1, Column.Gen2)]
 public class Benchmarks
 {
-    private const int ResultValue = 0;
-    private const string ErrorMessage = "An unknown error occured.";
-    private static readonly Error Error = new(ErrorMessage);
-    private static readonly Result FailedResult = Result.Fail(Error);
-    
     [Params(100_000)]
     public int Iterations { get; set; }
+
+    private const int ResultValue = 0;
+    private const string ErrorMessage = "An unknown error occured.";
+    private static readonly Error ErrorWithMessage = new(ErrorMessage);
+    private static readonly Result ResultOk = Result.Ok();
+    private static readonly Result ResultFail = Result.Fail();
+    private static readonly Result ResultFailWithMessage = Result.Fail(ErrorWithMessage);
+    private static readonly Result<int> ResultTValueOk = Result<int>.Ok(ResultValue);
+    private static readonly Result<int> ResultTValueFail = Result<int>.Fail();
+    private static readonly Result<int> ResultTValueFailWithMessage = Result<int>.Fail(ErrorWithMessage);
+    private static readonly CustomResult CustomResultOk = CustomResult.Ok();
+    private static readonly CustomResult CustomResultFail = CustomResult.Fail();
+    private static readonly CustomResult CustomResultFailWithMessage = CustomResult.Fail(ErrorWithMessage);
 
     [Benchmark]
     public void Develop_ResultBaseIndexer()
     {
         for (var iteration = 0; iteration < Iterations; iteration++)
-            _ = FailedResult.Errors[0];
+            _ = ResultFailWithMessage.Errors[0];
     }
 
     [Benchmark]
     public void Develop_ResultBaseHasError()
     {
         for (var iteration = 0; iteration < Iterations; iteration++)
-            _ = FailedResult.HasError<Error>();
+            _ = ResultFailWithMessage.HasError<Error>();
+    }
+
+    [Benchmark]
+    public void Develop_ResultOkToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = ResultOk.ToString();
+    }
+
+    [Benchmark]
+    public void Develop_ResultFailToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = ResultFail.ToString();
+    }
+
+    [Benchmark]
+    public void Develop_ResultFailWithMessageToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = ResultFailWithMessage.ToString();
+    }
+
+    [Benchmark]
+    public void Develop_ResultTValueOkToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = ResultTValueOk.ToString();
+    }
+
+    [Benchmark]
+    public void Develop_ResultTValueFailToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = ResultTValueFail.ToString();
+    }
+
+
+    [Benchmark]
+    public void Develop_ResultTValueFailWithMessageToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = ResultTValueFailWithMessage.ToString();
+    }
+
+    [Benchmark]
+    public void Develop_CustomResultOkToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = CustomResultOk.ToString();
+    }
+
+    [Benchmark]
+    public void Develop_CustomResultFailToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = CustomResultFail.ToString();
+    }
+
+    [Benchmark]
+    public void Develop_CustomResultFailWithMessageToString()
+    {
+        for (var iteration = 0; iteration < Iterations; iteration++)
+            _ = CustomResultFailWithMessage.ToString();
     }
 
     [Benchmark]
@@ -46,7 +122,7 @@ public class Benchmarks
     public void Develop_ResultFailWithError()
     {
         for (var iteration = 0; iteration < Iterations; iteration++)
-            _ = Result.Fail(Error);
+            _ = Result.Fail(ErrorWithMessage);
     }
 
     [Benchmark]
@@ -67,7 +143,7 @@ public class Benchmarks
     public void Develop_ResultTValueFailWithError()
     {
         for (var iteration = 0; iteration < Iterations; iteration++)
-            _ = Result<int>.Fail(Error);
+            _ = Result<int>.Fail(ErrorWithMessage);
     }
 
     [Benchmark]
@@ -89,7 +165,7 @@ public class Benchmarks
     public void Develop_ResultFailWithErrorTValue()
     {
         for (var iteration = 0; iteration < Iterations; iteration++)
-            _ = Result.Fail<int>(Error);
+            _ = Result.Fail<int>(ErrorWithMessage);
     }
 
     [Benchmark]
@@ -104,5 +180,31 @@ public class Benchmarks
     {
         for (var iteration = 0; iteration < Iterations; iteration++)
             _ = new Error(ErrorMessage);
+    }
+
+    private sealed class CustomResult : ResultBase
+    {
+        private CustomResult()
+        {
+        }
+
+        private CustomResult(IError error) : base(error)
+        {
+        }
+
+        public static CustomResult Ok()
+        {
+            return new CustomResult();
+        }
+
+        public static CustomResult Fail()
+        {
+            return new CustomResult(new Error());
+        }
+
+        public static CustomResult Fail(IError error)
+        {
+            return new CustomResult(error);
+        }
     }
 }